define basic structure of ComponentWithOverlay component
