+++ src/frontEnd/volunteering/SearchBar.tsx
@@ -22,6 +22,7 @@ export const SearchBar = (props) => {
         renderSuggestion,
         searchBarValue,
         onSearchBarValueChange,
+        onSearchBarGetsOrLosesFocus
     } 
     
     return <Page>
